Moving form to GroupAdd
